refactor: co-locate FilePaths with Workspace
Move FilePaths out of models.py into workspace.py, so the module owns both
run-scoped path types: FilePaths (the resolved tmp dir + its two JSON files)
and Workspace (the path calculator). FilePaths was standalone in models and
is the close sibling of Workspace (FilePaths.TMP_DIR == Workspace.tdir).

Also fixes a mypy finding in test_conversion surfaced by `just check`: the
media-info invariant test used `list.append(...) or []`; replaced with a
plain function.
